+++ src/route/image-router.ts
@@ -1,6 +1,11 @@
 import express from 'express';
+import { ImageController } from '../controller/image-controller';
+import { ImageMiddleware } from '../middleware/image-middleware';
 
-const imageRouter = (imageController, imageMiddleware) => {
+const imageRouter = (
+  imageController: ImageController,
+  imageMiddleware: ImageMiddleware,
+) => {
   const router = express.Router();
 
   router.post(
